requirements.txt
flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.0.1
httpx==0.28.0
notion-client==2.2.1
anthropic==0.40.0
